Refactor : Edit 폴더 Project로 이동, 불필요한 파일 삭제
