--- src/components/detail/index.ts
@@ -1,2 +1,0 @@
-export { default as DetailDiv } from './DetailDiv';
-export { default as ProjectAdd } from '../detail/ProjectAdd';
