+++ src/components/index.ts
@@ -1,6 +1,5 @@
-export { DetailDiv, ProjectAdd } from './detail';
 export { IconDiv, NavLink, UserPicture } from './nav';
-export { ProjectCard, ProjectDiv, ProjectEdit } from './project';
+export { ProjectCard, ProjectDiv, ProjectEdit, ProjectAdd, ProjectDetail } from './project';
 export { Skills, Summary } from './about-me';
 export { default as SubTitle } from './SubTitle';
 export { default as SubTitleEdit } from './SubTitleEdit';
