+++ src/pages/nav/ProjectDetailPage.tsx
@@ -1,4 +1,4 @@
-import { DetailDiv } from '../../components';
+import { ProjectDetail } from '../../components';
 import SubTitleEdit from '../../components/SubTitleEdit';
 import { useState } from 'react';
 
@@ -10,7 +10,7 @@ const ProjectDetailPage = () => {
     return (
         <>
             <SubTitleEdit text={`📂 Project`} onClick={handleChangeToggle} editMode={editToggle} />
-            <DetailDiv handleChangeToggle={handleChangeToggle} editMode={editToggle} />
+            <ProjectDetail handleChangeToggle={handleChangeToggle} editMode={editToggle} />
         </>
     );
 };
